--- go/lib/snet/snetproxy/conn_error_test.go
@@ -1,114 +1,0 @@
-// Copyright 2018 ETH Zurich
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package snetproxy_test
-
-import (
-	"net"
-	"os"
-	"syscall"
-	"testing"
-
-	"github.com/golang/mock/gomock"
-	. "github.com/smartystreets/goconvey/convey"
-
-	"github.com/scionproto/scion/go/lib/addr"
-	"github.com/scionproto/scion/go/lib/common"
-	"github.com/scionproto/scion/go/lib/snet/snetproxy"
-	"github.com/scionproto/scion/go/lib/snet/snetproxy/mock_snetproxy"
-)
-
-func TestNetworkFatalError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	Convey("Initialize proxy network", t, func() {
-		err := common.NewBasicError("Not dispatcher dead error, e.g., malformed register msg", nil)
-		mockNetwork := mock_snetproxy.NewMockNetwork(ctrl)
-		proxyNetwork := snetproxy.NewProxyNetwork(mockNetwork)
-		Convey("Dial error", func() {
-			mockNetwork.EXPECT().
-				DialSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any(), Any()).
-				Return(nil, err)
-			_, err := proxyNetwork.DialSCIONWithBindSVC("udp4", nil, nil, nil, addr.SvcNone, 0)
-			SoMsg("err", err, ShouldNotBeNil)
-		})
-		Convey("Listen error", func() {
-			mockNetwork.EXPECT().
-				ListenSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any()).
-				Return(nil, err)
-			_, err := proxyNetwork.ListenSCIONWithBindSVC("udp4", nil, nil, addr.SvcNone, 0)
-			SoMsg("err", err, ShouldNotBeNil)
-		})
-	})
-}
-
-func TestNetworkDispatcherDeadError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	dispatcherError := &net.OpError{Err: os.NewSyscallError("connect", syscall.ECONNREFUSED)}
-	Convey("Listen and Dial should reattempt to connect on dispatcher down errors", t, func() {
-		mockNetwork := mock_snetproxy.NewMockNetwork(ctrl)
-		proxyNetwork := snetproxy.NewProxyNetwork(mockNetwork)
-		Convey("Dial tries to reconnect if no timeout set", func() {
-			mockConn := NewMockConnWithAddrs(ctrl, localAddr, remoteAddr, nil, addr.SvcNone)
-			gomock.InOrder(
-				mockNetwork.EXPECT().
-					DialSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any(), Any()).
-					Return(nil, dispatcherError).
-					Times(2),
-				mockNetwork.EXPECT().
-					DialSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any(), Any()).
-					Return(mockConn, nil),
-			)
-			_, err := proxyNetwork.DialSCIONWithBindSVC("udp4", nil, nil, nil, addr.SvcNone, 0)
-			SoMsg("err", err, ShouldBeNil)
-		})
-		Convey("Dial only retries for limited time if timeout set", func() {
-			gomock.InOrder(
-				mockNetwork.EXPECT().
-					DialSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any(), Any()).
-					Return(nil, dispatcherError).
-					MinTimes(3).MaxTimes(5),
-			)
-			_, err := proxyNetwork.DialSCIONWithBindSVC("udp4",
-				nil, nil, nil, addr.SvcNone, tickerMultiplier(4))
-			SoMsg("err", err, ShouldNotBeNil)
-		})
-		Convey("Listen tries to reconnect if no timeout set", func() {
-			mockConn := NewMockConnWithAddrs(ctrl, localAddr, nil, nil, addr.SvcNone)
-			gomock.InOrder(
-				mockNetwork.EXPECT().
-					ListenSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any()).
-					Return(nil, dispatcherError).
-					Times(2),
-				mockNetwork.EXPECT().
-					ListenSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any()).
-					Return(mockConn, nil),
-			)
-			_, err := proxyNetwork.ListenSCIONWithBindSVC("udp4", nil, nil, addr.SvcNone, 0)
-			SoMsg("err", err, ShouldBeNil)
-		})
-		Convey("Listen only retries for limited time if timeout set", func() {
-			gomock.InOrder(
-				mockNetwork.EXPECT().
-					ListenSCIONWithBindSVC(Any(), Any(), Any(), Any(), Any()).
-					Return(nil, dispatcherError).
-					MinTimes(3).MaxTimes(5),
-			)
-			_, err := proxyNetwork.ListenSCIONWithBindSVC("udp4",
-				nil, nil, addr.SvcNone, tickerMultiplier(4))
-			SoMsg("err", err, ShouldNotBeNil)
-		})
-	})
-}
